+++ packages/ai/core/generate-object/output-strategy.ts
@@ -14,7 +14,7 @@ import {
   ValidationResult,
 } from '@ai-sdk/provider-utils';
 import * as z3 from 'zod/v3';
-import * as z4 from 'zod/v4/core';
+import * as z4 from 'zod/v4';
 import { NoObjectGeneratedError } from '../../src/error/no-object-generated-error';
 import {
   AsyncIterableStream,
@@ -394,7 +394,7 @@ export function getOutputStrategy<SCHEMA>({
 }: {
   output: 'object' | 'array' | 'enum' | 'no-schema';
   schema?:
-    | z4.$ZodType<SCHEMA, any>
+    | z4.ZodType<SCHEMA, any>
     | z3.Schema<SCHEMA, z3.ZodTypeDef, any>
     | Schema<SCHEMA>;
   enumValues?: Array<SCHEMA>;