+++ packages/valibot/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/valibot
 
+## 1.0.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [0571b98]
+- Updated dependencies [39a4fab]
+- Updated dependencies [d1a034f]
+  - @ai-sdk/provider-utils@3.0.0-beta.2
+
 ## 1.0.0-beta.1
 
 ### Patch Changes